+++ lilith-data/logging/src/test/groovy/de/huxhorn/lilith/data/logging/ExtendedStackTraceElementSpec.groovy
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2013 Joern Huxhorn
+ * Copyright (C) 2007-2016 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,7 +17,7 @@
  */
 
 /*
- * Copyright 2007-2013 Joern Huxhorn
+ * Copyright 2007-2016 Joern Huxhorn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -319,8 +319,8 @@ class ExtendedStackTraceElementSpec extends Specification {
             'className.methodName(Unknown Source)',
             'className.methodName(fileName)',
             'className.methodName(Unknown Source)',
-            'className.methodName(Unknown Source) ~[codeLocation:]',
-            'className.methodName(Unknown Source) ~[:version]',
+            'className.methodName(Unknown Source) ~[codeLocation:na]',
+            'className.methodName(Unknown Source) ~[na:version]',
             'className.methodName(Unknown Source)',
             'className.methodName(fileName:17) [codeLocation:version]',
             'className.methodName(Native Method) [codeLocation:version]',