+++ CHANGELOG.md
@@ -24,6 +24,9 @@ All notable changes to this project will be documented in this file.
 - Changed specification of HTTP status code 451 from draft to RFC 7725.
 - Changed meaning of `ThrowableCondition`. Condition now evaluates to true if search string is one of the `Throwable` contained in the hierarchy (including cause, suppressed) instead of only checking the root exception.
 - `EventContainsCondition` is now evaluating `Throwable` and NDC.
+- `appendTo` methods of `ThrowableInfo` and `ExtendedStackTraceElement` now throw a `NullPointerException` if `StringBuilder` is `null`.
+- `ExtendedStackTraceElement.appendExtended` is now private.
+- `ExtendedStackTraceElement.toString(true)` is now printing `"na"` instead of empty string if codeLocation or version is null. 
 
 ### Deprecated
 - Nothing.